+++ src/app/_components/SpotInfoWindow.tsx
