+++ src/app/_components/SpotMarker.tsx
